Update DB Primary Keys

+++ FarmFresh/FarmFresh.Framework/Entities/Categories/ProductCategory.cs
@@ -3,13 +3,13 @@ using FarmFresh.Framework.Entities.Products;
 
 namespace FarmFresh.Framework.Entities.Categories
 {
-    public class ProductCategory : IEntity<int>
+    public class ProductCategory : IEntity<long>
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public virtual Product Product { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
 
         public virtual Category Category { get; set; }
     }
