+++ FarmFresh/FarmFresh.Framework/Entities/Stores/ProductStore.cs
@@ -3,13 +3,13 @@ using FarmFresh.Framework.Entities.Products;
 
 namespace FarmFresh.Framework.Entities.Stores
 {
-    public class ProductStore : IEntity<int>
+    public class ProductStore : IEntity<long>
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public virtual Product Product { get; set; }
 
-        public int StoreId { get; set; }
+        public long StoreId { get; set; }
 
         public virtual Store Store { get; set; }
     }
