+++ FarmFresh/FarmFresh.Framework/Repositories/Abstract/IOrderRepository.cs
@@ -4,7 +4,7 @@ using FarmFresh.Framework.Entities.Orders;
 
 namespace FarmFresh.Framework.Repositories.Abstract
 {
-    public interface IOrderRepository : IRepository<Order, int, FrameworkContext>
+    public interface IOrderRepository : IRepository<Order, long, FrameworkContext>
     {
     }
 }
